misc: Rename fast xml http module to use kebab case naming
