+++ src/gmp/commands/oci-image-targets.ts
@@ -5,7 +5,7 @@
 
 import EntitiesCommand from 'gmp/commands/entities';
 import type Http from 'gmp/http/http';
-import {type XmlResponseData} from 'gmp/http/transform/fastxml';
+import {type XmlResponseData} from 'gmp/http/transform/fast-xml';
 import OciImageTarget from 'gmp/models/oci-image-target';
 
 class OciImageTargetsCommand extends EntitiesCommand<OciImageTarget> {
