+++ src/gmp/commands/role.ts
@@ -5,7 +5,7 @@
 
 import EntityCommand from 'gmp/commands/entity';
 import type Http from 'gmp/http/http';
-import {type XmlResponseData} from 'gmp/http/transform/fastxml';
+import {type XmlResponseData} from 'gmp/http/transform/fast-xml';
 import logger from 'gmp/log';
 import Role, {type RoleElement} from 'gmp/models/role';
 import {isArray} from 'gmp/utils/identity';
