+++ src/web/entity/hooks/useEntityClone.ts
@@ -6,7 +6,7 @@
 import {type EntityActionData} from 'gmp/commands/entity';
 import type Rejection from 'gmp/http/rejection';
 import type Response from 'gmp/http/response';
-import {type XmlMeta} from 'gmp/http/transform/fastxml';
+import {type XmlMeta} from 'gmp/http/transform/fast-xml';
 import {type EntityType} from 'gmp/utils/entity-type';
 import actionFunction from 'web/entity/hooks/action-function';
 import useGmp from 'web/hooks/useGmp';
