+++ src/web/hooks/use-query/agents.ts
@@ -7,7 +7,7 @@ import {useQueryClient} from '@tanstack/react-query';
 import {type AgentModifyParams} from 'gmp/commands/agent';
 import type Rejection from 'gmp/http/rejection';
 import type Response from 'gmp/http/response';
-import {type XmlMeta} from 'gmp/http/transform/fastxml';
+import {type XmlMeta} from 'gmp/http/transform/fast-xml';
 import type Agent from 'gmp/models/agent';
 import Filter from 'gmp/models/filter';
 import {isFilter} from 'gmp/models/filter/utils';
