+++ src/web/hooks/use-query/oci-image-targets.ts
@@ -10,7 +10,7 @@ import {
 } from 'gmp/commands/oci-image-target';
 import type Rejection from 'gmp/http/rejection';
 import type Response from 'gmp/http/response';
-import {type XmlMeta} from 'gmp/http/transform/fastxml';
+import {type XmlMeta} from 'gmp/http/transform/fast-xml';
 import type Filter from 'gmp/models/filter';
 import {isFilter} from 'gmp/models/filter/utils';
 import type OciImageTarget from 'gmp/models/oci-image-target';
