+++ src/web/pages/agent-groups/AgentGroupsComponent.tsx
@@ -7,7 +7,7 @@ import React, {useState} from 'react';
 import {type EntityActionData} from 'gmp/commands/entity';
 import type Rejection from 'gmp/http/rejection';
 import type Response from 'gmp/http/response';
-import {type XmlMeta} from 'gmp/http/transform/fastxml';
+import {type XmlMeta} from 'gmp/http/transform/fast-xml';
 import type AgentGroup from 'gmp/models/agent-group';
 import {isDefined} from 'gmp/utils/identity';
 import useEntityClone from 'web/entity/hooks/useEntityClone';
