+++ src/web/pages/scanners/ScannerComponent.tsx
@@ -6,7 +6,7 @@
 import {useState} from 'react';
 import {type EntityActionData} from 'gmp/commands/entity';
 import type Response from 'gmp/http/response';
-import {type XmlMeta} from 'gmp/http/transform/fastxml';
+import {type XmlMeta} from 'gmp/http/transform/fast-xml';
 import type ActionResult from 'gmp/models/action-result';
 import {
   type default as Credential,
